Resolved issue where blueprint was not registering.
