--- backend/api/__init__.py
@@ -1,8 +1,0 @@
-from flask import Flask
-from search import search
-
-def create_app():
-    app = Flask(__name__)
-    app.register_blueprint(search)  
-
-    return app
