+++ backend/api/search.py
@@ -1,8 +1,8 @@
 from flask import Blueprint
 
 # Create a blueprint instance
-search_bp = Blueprint('search', __name__, url_prefix='/search')
+search_bp = Blueprint('search', __name__)
 
-@search_bp.route('/search')
-def get_locations():
+@search_bp.route('/')
+def get_users():
     return {"message": "List of users"}
